requirements.txt
# NextFactory ERP+MES Exhibition Demo - Python Dependencies
# Phase 1: Foundation Setup Dependencies

# Core PyQt6 GUI Framework
PyQt6==6.6.1
PyQt6-Qt6==6.6.1

# Database ORM and PostgreSQL adapter
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# Data analysis and visualization for reporting modules
pandas==2.1.4
matplotlib==3.8.2
plotly==5.18.0

# Date and time utilities
python-dateutil==2.8.2

# Configuration management
python-dotenv==1.0.0

# Utility libraries
bcrypt==4.1.2
typing-extensions==4.9.0

# Development and testing (optional for exhibition)
pytest==7.4.4
pytest-qt==4.3.1

# Package building for exhibition deployment
pyinstaller==6.3.0